tg-pdf-secure/requirements.txt
python-telegram-bot>=21.9,<22.0
httpx>=0.27.2,<0.28.0
python-dotenv>=1.0.1,<2.0.0
